refactor: move gas limit tests to tests/gas_limit_test.rs
